:truck: Moved interfaces and classes to folders
