--- lib/collection.ts
@@ -1,5 +1,0 @@
-import { LearningObject } from './learning-object';
-export type Collection = {
-    name: string;
-    learningObjects: LearningObject[];
-}
