+++ lib/collection/collection.ts
@@ -1,0 +1,5 @@
+import { LearningObject } from '../learning-object/learning-object';
+export type Collection = {
+  name: string;
+  learningObjects: LearningObject[];
+};
