+++ lib/index.ts
@@ -1,13 +1,16 @@
-import { User } from './user';
-import { LearningObject, AcademicLevel } from './learning-object';
-import { Outcome, OutcomeSuggestion } from './outcome';
-import { StandardOutcome } from './standard-outcome';
-import { LearningOutcome } from './learning-outcome';
+import { User } from './user/user';
+import {
+  LearningObject,
+  AcademicLevel
+} from './learning-object/learning-object';
+import { Outcome, OutcomeSuggestion } from './outcome/outcome';
+import { StandardOutcome } from './standard-outcome/standard-outcome';
+import { LearningOutcome } from './learning-outcome/learning-outcome';
 import { LearningGoal } from './learning-goal';
 import { InstructionalStrategy } from './instructional-strategy';
 import { AssessmentPlan } from './assessment-plan';
-import { Collection } from './collection';
-import { Rating } from './rating';
+import { Collection } from './collection/collection';
+import { Rating } from './rating/rating';
 
 //Exports All Interfaces
 export {
