+++ lib/outcome/outcome.ts
