+++ lib/rating/rating.ts
@@ -1,4 +1,4 @@
-import { User } from './user';
+import { User } from '../user/user';
 
 export interface Rating {
   id?: string;
